Move createVideoPlayerViewModel
