+++ Shared/ViewModels/ItemViewModel.swift
@@ -36,7 +36,7 @@ class ItemViewModel: ViewModel {
 
         getSimilarItems()
         
-        self.createVideoPlayerViewModel(item: item)
+        item.createVideoPlayerViewModel()
             .sink { completion in
                 self.handleAPIRequestError(completion: completion)
             } receiveValue: { videoPlayerViewModel in
@@ -110,97 +110,5 @@ class ItemViewModel: ViewModel {
                 })
                 .store(in: &cancellables)
         }
-    }
-    
-    func createVideoPlayerViewModel(item: BaseItemDto) -> AnyPublisher<VideoPlayerViewModel, Error> {
-        let builder = DeviceProfileBuilder()
-        // TODO: fix bitrate settings
-        builder.setMaxBitrate(bitrate: 60000000)
-        let profile = builder.buildProfile()
-        
-        let playbackInfo = PlaybackInfoDto(userId: SessionManager.main.currentLogin.user.id,
-                                           maxStreamingBitrate: 60000000,
-                                           startTimeTicks: item.userData?.playbackPositionTicks ?? 0,
-                                           deviceProfile: profile,
-                                           autoOpenLiveStream: true)
-        
-        return MediaInfoAPI.getPostedPlaybackInfo(itemId: item.id!,
-                                           userId: SessionManager.main.currentLogin.user.id,
-                                           maxStreamingBitrate: 60000000,
-                                           startTimeTicks: item.userData?.playbackPositionTicks ?? 0,
-                                           autoOpenLiveStream: true,
-                                           playbackInfoDto: playbackInfo)
-            .map({ response -> VideoPlayerViewModel in
-                let mediaSource = response.mediaSources!.first!
-                
-                let audioStreams = mediaSource.mediaStreams?.filter({ $0.type == .audio }) ?? []
-                let subtitleStreams = mediaSource.mediaStreams?.filter({ $0.type == .subtitle }) ?? []
-                
-                let defaultAudioStream = audioStreams.first(where: { $0.index! == mediaSource.defaultAudioStreamIndex! })
-                
-                let defaultSubtitleStream = subtitleStreams.first(where: { $0.index! == mediaSource.defaultSubtitleStreamIndex ?? -1 })
-                
-                let videoStream = mediaSource.mediaStreams!.first(where: { $0.type! == MediaStreamType.video })
-                
-                let audioCodecs = mediaSource.mediaStreams!.filter({ $0.type! == MediaStreamType.audio }).map({ $0.codec! })
-                
-                // MARK: basic stream
-                var streamURL = URLComponents(string: SessionManager.main.currentLogin.server.currentURI)!
-                streamURL.path = "/Videos/\(item.id!)/stream"
-
-                streamURL.addQueryItem(name: "Static", value: "true")
-                streamURL.addQueryItem(name: "MediaSourceId", value: item.id!)
-                streamURL.addQueryItem(name: "Tag", value: item.etag)
-                streamURL.addQueryItem(name: "MinSegments", value: "6")
-                
-                // MARK: hls stream
-                var hlsURL = URLComponents(string: SessionManager.main.currentLogin.server.currentURI)!
-                hlsURL.path = "/videos/\(item.id!)/master.m3u8"
-
-                hlsURL.addQueryItem(name: "DeviceId", value: UIDevice.vendorUUIDString)
-                hlsURL.addQueryItem(name: "MediaSourceId", value: item.id!)
-                hlsURL.addQueryItem(name: "VideoCodec", value: videoStream?.codec!)
-                hlsURL.addQueryItem(name: "AudioCodec", value: audioCodecs.joined(separator: ","))
-                hlsURL.addQueryItem(name: "AudioStreamIndex", value: "\(defaultAudioStream!.index!)")
-                hlsURL.addQueryItem(name: "VideoBitrate", value: "\(videoStream!.bitRate!)")
-                hlsURL.addQueryItem(name: "AudioBitrate", value: "\(defaultAudioStream!.bitRate!)")
-                hlsURL.addQueryItem(name: "PlaySessionId", value: response.playSessionId!)
-                hlsURL.addQueryItem(name: "TranscodingMaxAudioChannels", value: "6")
-                hlsURL.addQueryItem(name: "RequireAvc", value: "false")
-                hlsURL.addQueryItem(name: "Tag", value: mediaSource.eTag!)
-                hlsURL.addQueryItem(name: "SegmentContainer", value: "ts")
-                hlsURL.addQueryItem(name: "MinSegments", value: "2")
-                hlsURL.addQueryItem(name: "BreakOnNonKeyFrames", value: "true")
-                hlsURL.addQueryItem(name: "TranscodeReasons", value: "VideoCodecNotSupported,AudioCodecNotSupported")
-                hlsURL.addQueryItem(name: "api_key", value: SessionManager.main.currentLogin.user.accessToken)
-                
-                if defaultSubtitleStream?.index != nil {
-                    hlsURL.addQueryItem(name: "SubtitleMethod", value: "Encode")
-                    hlsURL.addQueryItem(name: "SubtitleStreamIndex", value: "\(defaultSubtitleStream!.index!)")
-                }
-
-//                    startURL.queryItems?.append(URLQueryItem(name: "SubtitleCodec", value: "\(defaultSubtitleStream!.codec!)"))
-                
-                let videoPlayerViewModel = VideoPlayerViewModel(item: item,
-                                                                title: item.name!,
-                                                                subtitle: item.seriesName,
-                                                                streamURL: streamURL.url!,
-                                                                hlsURL: hlsURL.url!,
-                                                                response: response,
-                                                                audioStreams: audioStreams,
-                                                                subtitleStreams: subtitleStreams,
-                                                                defaultAudioStreamIndex: defaultAudioStream?.index ?? -1,
-                                                                defaultSubtitleStreamIndex: defaultSubtitleStream?.index ?? -1,
-                                                                playerState: .playing,
-                                                                shouldShowGoogleCast: false,
-                                                                shouldShowAirplay: false,
-                                                                subtitlesEnabled: defaultAudioStream?.index != nil,
-                                                                sliderPercentage: (item.userData?.playedPercentage ?? 0) / 100,
-                                                                selectedAudioStreamIndex: defaultAudioStream?.index ?? -1,
-                                                                selectedSubtitleStreamIndex: defaultSubtitleStream?.index ?? -1)
-                
-                return videoPlayerViewModel
-            })
-            .eraseToAnyPublisher()
     }
 }
